requirements.txt
numpy>=1.25.2,<2
matplotlib>=3.6.2
scipy>=1.9.3,<2
tqdm>=4.64.1
tqdm_joblib>=0.0.4
pyPCG_toolbox>=0.1b5,<1
joblib>=1.2.0,<2